Split TypeChecker statement inference into topic modules
